# docs/commercialization/openapi/open-v1.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "YINYU CTF Platform Open API",
    "description": "YINYU CTF Platform external API contract",
    "version": "v1"
  },
  "paths": {
...
    "/api/open/v1/bootstrap-profiles": {
      "post": {
        "tags": [
          "OpenBootstrapProfiles"
        ],
        "operationId": "OpenBootstrapProfiles_Create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "x-originalName": "idempotencyKey",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapProfileCreateModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOperationModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      },
      "get": {
        "tags": [
          "OpenBootstrapProfiles"
        ],
        "operationId": "OpenBootstrapProfiles_List",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 100.0,
              "minimum": 1.0
            },
            "x-position": 1
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapProfileCursorPage"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
    "/api/open/v1/bootstrap-profiles/{profileId}": {
      "get": {
        "tags": [
          "OpenBootstrapProfiles"
        ],
        "operationId": "OpenBootstrapProfiles_Get",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "OpenBootstrapProfiles"
        ],
        "operationId": "OpenBootstrapProfiles_Delete",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "Idempotency-Key",
            "x-originalName": "idempotencyKey",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOperationModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
    "/api/open/v1/bootstrap-profiles/{profileId}/versions": {
      "post": {
        "tags": [
          "OpenBootstrapProfiles"
        ],
        "operationId": "OpenBootstrapProfiles_PublishVersion",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "Idempotency-Key",
            "x-originalName": "idempotencyKey",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "artifact": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  },
                  "manifest": {
                    "type": "string",
                    "nullable": true
                  },
                  "version": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "expectedDigest": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOperationModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
    "/api/open/v1/bootstrap-profiles/{profileId}/versions/{version}": {
      "get": {
        "tags": [
          "OpenBootstrapProfiles"
        ],
        "operationId": "OpenBootstrapProfiles_GetVersion",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
...
    "/api/open/v1/operations/{id}": {
      "get": {
        "tags": [
          "Operations"
        ],
        "operationId": "Operations_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOperationModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ApiOperationModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "kind": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiOperationStatus"
          },
          "stage": {
            "type": "string"
          },
          "resourceType": {
            "type": "string",
            "nullable": true
          },
          "resourceId": {
            "type": "string",
            "nullable": true
          },
          "deploymentQueueTicketId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "currentProgress": {
            "type": "integer",
            "format": "int64"
          },
          "totalProgress": {
            "type": "integer",
            "format": "int64"
          },
          "attemptCount": {
            "type": "integer",
            "format": "int32"
          },
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "errorDetail": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "nullable": true
          },
          "createdAt": {
            "type": "integer",
            "format": "uint64"
          },
          "startedAt": {
            "type": "integer",
            "format": "uint64",
            "nullable": true
          },
          "updatedAt": {
            "type": "integer",
            "format": "uint64"
          },
          "completedAt": {
            "type": "integer",
            "format": "uint64",
            "nullable": true
          }
        }
      },
      "ApiOperationStatus": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Pending",
          "Running",
          "Succeeded",
          "Failed"
        ],
        "enum": [
          0,
          1,
          2,
          3
        ]
      },
...
      "BootstrapProfileCreateModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BootstrapProfileCursorPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BootstrapProfileModel"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BootstrapProfileModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/BootstrapProfileStatus"
          },
          "latestVersion": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "createdAt": {
            "type": "integer",
            "format": "uint64"
          },
          "updatedAt": {
            "type": "integer",
            "format": "uint64",
            "nullable": true
          }
        }
      },
      "BootstrapProfileStatus": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Active",
          "Deleting",
          "Deleted"
        ],
        "enum": [
          0,
          1,
          2
        ]
      },
...
    },
    "securitySchemes": {
      "GzctfApiToken": {
        "type": "http",
        "description": "Scoped GZCTF API token",
        "scheme": "bearer",
        "bearerFormat": "GZCTF API token"
      }
    }
  },
...
}